Refactored chat package to reside in parent package. Renamed a few files to better represent their functionality.
