+++ src/main/java/catchat/data/source/GroupMeGroupDS.java
@@ -1,8 +1,8 @@
-package catchat.data.source.chats;
+package catchat.data.source;
 
 import catchat.data.entities.chat.Chat;
-import catchat.data.entities.chat.DirectChat;
-import catchat.data.entities.message.DirectMessage;
+import catchat.data.entities.chat.GroupChat;
+import catchat.data.entities.message.GroupMessage;
 import catchat.data.entities.message.Message;
 import catchat.data.entities.profile.MemberProfile;
 import catchat.data.entities.profile.Profile;
@@ -16,19 +16,19 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by andrew on 4/17/18.
+ * Created by andrew on 4/13/18.
  */
-public class GroupMeDirectChatDS extends ChatDataSource {
-    private static GroupMeDirectChatDS INSTANCE;
+public class GroupMeGroupDS extends DataSource {
+    private static GroupMeGroupDS INSTANCE;
     private ObjectMapper mapper;
 
-    private GroupMeDirectChatDS() {
+    private GroupMeGroupDS() {
         mapper = new ObjectMapper();
     }
 
-    public static GroupMeDirectChatDS getInstance() {
+    public static GroupMeGroupDS getInstance() {
         if (INSTANCE == null) {
-            INSTANCE = new GroupMeDirectChatDS();
+            INSTANCE = new GroupMeGroupDS();
         }
         return INSTANCE;
     }
@@ -36,7 +36,7 @@ public class GroupMeDirectChatDS extends ChatDataSource {
     @Override
     public void getChats(int page, int pageSize, ChatsCallback callback) {
         HttpRequestFactory httpRequestFactory = new NetHttpTransport().createRequestFactory();
-        GenericUrl url = new GenericUrl(BASE_API_URL + "chats");
+        GenericUrl url = new GenericUrl(BASE_API_URL + "groups");
         url.set("token", getAuthToken());
         url.set("page", page);
         url.set("per_page", pageSize);
@@ -51,7 +51,7 @@ public class GroupMeDirectChatDS extends ChatDataSource {
             switch (responseCode) {
                 case RESPONSE_OK :
                     JsonNode responseNode = responseTree.get("response");
-                    callback.onChatsLoaded(parseChats(responseNode));
+                    callback.onChatsLoaded(parseGroups(responseNode));
                     break;
                 default :
                     callback.unknownResponseCode(responseCode);
@@ -64,9 +64,8 @@ public class GroupMeDirectChatDS extends ChatDataSource {
     @Override
     public void getMessages(String chatId, String beforeMessageId, String sinceMessageId, MessagesCallback callback) {
         HttpRequestFactory httpRequestFactory = new NetHttpTransport().createRequestFactory();
-        GenericUrl url = new GenericUrl(BASE_API_URL + "direct_messages");
+        GenericUrl url = new GenericUrl(BASE_API_URL + "groups/" + chatId + "/messages");
         url.set("token", getAuthToken());
-        url.set("other_user_id", chatId);
 
         try {
             HttpRequest httpRequest = httpRequestFactory.buildGetRequest(url);
@@ -91,7 +90,7 @@ public class GroupMeDirectChatDS extends ChatDataSource {
     @Override
     public void sendMessage(String chatId, String sourceGUID, String messageText, MessagesCallback callback) {
         HttpRequestFactory httpRequestFactory = new NetHttpTransport().createRequestFactory();
-        GenericUrl url = new GenericUrl(BASE_API_URL + "direct_messages");
+        GenericUrl url = new GenericUrl(BASE_API_URL + "groups/" + chatId + "/messages");
         url.set("token", getAuthToken());
 
         sourceGUID = BASE_SOURCE_GUID + sourceGUID;
@@ -127,45 +126,51 @@ public class GroupMeDirectChatDS extends ChatDataSource {
 
     }
 
-    private List<Chat> parseChats(JsonNode responseNode) {
-        List<Chat> chatList = new ArrayList<>();
+    private List<Chat> parseGroups(JsonNode responseNode) {
+        List<Chat> groupList = new ArrayList<>();
+
         if (responseNode.isArray()) {
             for (JsonNode node : responseNode) {
+                String groupId = node.get("group_id").asText();
+                String name = node.get("name").asText();
+                String preview = node.get("messages").get("preview").get("text").asText();
+                JsonNode members = node.get("members");
+
                 List<Profile> memberList = new ArrayList<>();
-                String preview = node.get("last_message").get("text").asText();
-
-                JsonNode otherMember = node.get("other_user");
-                String otherName = otherMember.get("name").asText();
-                String otherId = otherMember.get("id").asText();
-                memberList.add(new MemberProfile(otherId, otherName, ""));
-
-                chatList.add(new DirectChat(otherId, otherName, preview, memberList));
+                if (members.isArray()) {
+                    for (JsonNode member : members) {
+                        String nickname = member.get("nickname").asText();
+                        String userId = member.get("user_id").asText();
+                        String memberId = member.get("id").asText();
+                        memberList.add(new MemberProfile(userId, nickname, memberId));
+                    }
+                }
+                groupList.add(new GroupChat(groupId, name, preview, memberList));
             }
         }
-        return chatList;
+        return groupList;
     }
 
     private List<Message> parseMessages(JsonNode responseNode) {
         List<Message> messageList = new ArrayList<>();
-        JsonNode messages = responseNode.get("direct_messages");
 
+        JsonNode messages = responseNode.get("messages");
         if (messages.isArray()) {
             for (JsonNode message : messages) {
                 String messageId = message.get("id").asText();
                 String messageGUID = message.get("source_guid").asText();
                 String senderId = message.get("sender_id").asText();
-                long createdAt = message.get("created_at").asLong();
                 String text = message.get("text").asText();
-                messageList.add(new DirectMessage(messageId, messageGUID, text, senderId, createdAt));
+                long createdAt = message.get("created_at").asLong();
+                messageList.add(new GroupMessage(messageId, messageGUID, text, senderId, createdAt));
             }
         }
         return messageList;
     }
 
     private byte[] createMessage(String chatId, String sourceGUID, String messageText) {
-        String message = "{\"direct_message\": {";
+        String message = "{\"message\": {";
         message += "\"source_guid\": \"" + sourceGUID + "\", ";
-        message += "\"recipient_id\": \"" + chatId + "\", ";
         message += "\"text\": \"" + messageText + "\"}}";
         try {
             return message.getBytes("UTF-8");
